backend/requirements.txt
Django>=5.0
djangorestframework
dj-rest-auth
django-allauth
django-cors-headers
psycopg2-binary
Pillow
python-decouple
djangorestframework-simplejwt
dj-rest-auth[with-social]
python-dotenv
django-allauth-ui
cloudinary 
django-cloudinary-storage
channels
channels_redis
daphne
django-taggit
typesense

